# static/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkiResort",
    "description": "API for Web Course Project",
    "version": "v1"
  },
  "paths": {
...
    "/cardReadings": {
      "get": {
        "tags": [
          "cardReadings"
        ],
        "summary": "Get information about all cardReadings",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Information about all cardReadings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardReading"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cardReadings"
        ],
        "summary": "Add a new cardReading",
        "parameters": [
          {
            "name": "turnstileID",
            "in": "query",
            "description": "ID of the turnstile where the reading took place",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cardID",
            "in": "query",
            "description": "ID of the card that was read",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "readingTime",
            "in": "query",
            "description": "The time of the card reading",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "The added cardReading with assigned ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardReading"
                }
              }
            }
          }
        }
      }
    },
    "/cardReadings/{recordID}": {
      "get": {
        "tags": [
          "cardReadings"
        ],
        "summary": "Get information about a cardReading by it's ID",
        "parameters": [
          {
            "name": "recordID",
            "in": "path",
            "description": "CardReading ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "CardReading with specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardReading"
                }
              }
            }
          },
          "404": {
            "description": "CardReading with specified ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cardReadings"
        ],
        "summary": "Update information about an existing cardReading",
        "parameters": [
          {
            "name": "recordID",
            "in": "path",
            "description": "ID of the cardReading to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "turnstileID",
            "in": "query",
            "description": "ID of the turnstile where the reading took place",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cardID",
            "in": "query",
            "description": "ID of the card that was read",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "readingTime",
            "in": "query",
            "description": "The time of the card reading",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The cardReading was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "A cardReading with specified ID was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cardReadings"
        ],
        "summary": "Delete a cardReading by it's ID",
        "parameters": [
          {
            "name": "recordID",
            "in": "path",
            "description": "ID of the cardReading to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "CardReading was successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "CardReading with specified ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/cards": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get information about all cards",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Information about all cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Add a new card",
        "parameters": [
          {
            "name": "activationTime",
            "in": "query",
            "description": "Time when the new card was activated",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "The type of the new card",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "The added card with assigned ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{cardID}": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get information about a card by it's ID",
        "parameters": [
          {
            "name": "cardID",
            "in": "path",
            "description": "Card ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Card with specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "404": {
            "description": "Card with specified ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cards"
        ],
        "summary": "Update information about an existing card",
        "parameters": [
          {
            "name": "cardID",
            "in": "path",
            "description": "ID of the card to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "activationTime",
            "in": "query",
            "description": "Time when the new card was activated",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "The type of the card",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The card was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "A card with specified ID was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cards"
        ],
        "summary": "Delete a card by it's ID",
        "parameters": [
          {
            "name": "cardID",
            "in": "path",
            "description": "ID of the card to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Card was successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "Card with specified ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Card": {
        "type": "object",
        "properties": {
          "cardID": {
            "type": "integer",
            "description": "Card ID",
            "format": "int32"
          },
          "activationTime": {
            "type": "string",
            "description": "The time when the card was activated",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "description": "Type of the card",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Information about cards used in the resort"
      },
      "CardReading": {
        "type": "object",
        "properties": {
          "recordID": {
            "type": "integer",
            "description": "The ID of the record about card reading",
            "format": "int32"
          },
          "turnstileID": {
            "type": "integer",
            "description": "ID of the turnstile where the reading took place",
            "format": "int32"
          },
          "cardID": {
            "type": "integer",
            "description": "ID of the card that was read",
            "format": "int32"
          },
          "readingTime": {
            "type": "string",
            "description": "The time of the card reading",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Information about readings of cards on turnstiles of the resort"
      },
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}